refactor: @utoo/web code styles
